requirements.txt
numpy
PyYAML
tensorboard
torch
